requirements.txt
python-telegram-bot==22.6
bcrypt==4.1.2
python-dotenv>=1.0.0
